Extract debounce function to utils
