+++ src/js/components/Browser.react.js
@@ -8,23 +8,7 @@
 
 let React = require('react');
 let fs = require('fs');
-
-// TODO: extract to utils
-function debounce(func, wait, immediate = false) {
-  var timeout;
-  return function() {
-    var context = this,
-      args = arguments;
-    var later = function() {
-      timeout = null;
-      if (!immediate) {func.apply(context, args);}
-    };
-    var callNow = immediate && !timeout;
-    clearTimeout(timeout);
-    timeout = setTimeout(later, wait);
-    if (callNow) {func.apply(context, args);}
-  };
-}
+let debounce = require('../utils/debounce.js');
 
 const homeURL = `file:///${__dirname}/../../html/home.html`;
